TSR-32 Added Maps, Changed Camera, Added AI model
This commit features a lot of changes. I have implemented the use of StreetMaps to show the location of the user and track him. Changed the camera functionality so now the app only takes photos based on the users speed while driving. Added our own custom AI model and tested it on a picture inside the app.